better_context/pyproject.toml
[project]
name = "wkend-projects"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chonkie[all]>=1.4.0",
    "crawl4ai>=0.7.4",
    "mcp>=1.17.0",
    "weaviate-client>=4.17.0",
]
